+++ test/query_test.exs
@@ -317,14 +317,18 @@ defmodule QueryTest do
 
   @tag min_pg_version: "9.0"
   test "decode network types", context do
-    assert [[%Postgrex.INET{address: {127, 0, 0, 1}}]] =
-           query("SELECT '127.0.0.1'::inet", [])
-    assert [[%Postgrex.INET{address: {8193, 43981, 0, 0, 0, 0, 0, 0}}]] =
-           query("SELECT '2001:abcd::'::inet", [])
-    assert [[%Postgrex.CIDR{address: {127, 0, 0, 1}, netmask: 32}]] =
+    assert [[%Postgrex.INET{address: {127, 0, 0, 1}, netmask: 32}]] =
+           query("SELECT '127.0.0.1/32'::inet", [])
+    assert [[%Postgrex.INET{address: {8193, 43981, 0, 0, 0, 0, 0, 0}, netmask: 128}]] =
+           query("SELECT '2001:abcd::/128'::inet", [])
+    assert [[%Postgrex.INET{address: {127, 0, 0, 1}, netmask: 8}]] =
+           query("SELECT '127.0.0.1/8'::inet", [])
+    assert [[%Postgrex.INET{address: {127, 0, 0, 1}, netmask: 32}]] =
            query("SELECT '127.0.0.1/32'::cidr", [])
-    assert [[%Postgrex.CIDR{address: {8193, 43981, 0, 0, 0, 0, 0, 0}, netmask: 128}]] =
+    assert [[%Postgrex.INET{address: {8193, 43981, 0, 0, 0, 0, 0, 0}, netmask: 128}]] =
            query("SELECT '2001:abcd::/128'::cidr", [])
+    assert [[%Postgrex.INET{address: {192, 168, 0, 0}, netmask: 16}]] =
+           query("SELECT '192.168.0.0/16'::cidr", [])
     assert [[%Postgrex.MACADDR{address: {8, 1, 43, 5, 7, 9}}]] =
            query("SELECT '08:01:2b:05:07:09'::macaddr", [])
   end
@@ -640,16 +644,23 @@ defmodule QueryTest do
 
   @tag min_pg_version: "9.0"
   test "encode network types", context do
-    assert [[%Postgrex.INET{address: {127, 0, 0, 1}}]] =
-           query("SELECT $1::inet", [%Postgrex.INET{address: {127, 0, 0, 1}}])
-    assert [[%Postgrex.INET{address: {8193, 43981, 0, 0, 0, 0, 0, 0}}]] =
-           query("SELECT $1::inet", [%Postgrex.INET{address: {8193, 43981, 0, 0, 0, 0, 0, 0}}])
-    assert [[%Postgrex.CIDR{address: {127, 0, 0, 1}, netmask: 32}]] =
-           query("SELECT $1::cidr", [%Postgrex.CIDR{address: {127, 0, 0, 1}, netmask: 32}])
-    assert [[%Postgrex.CIDR{address: {8193, 43981, 0, 0, 0, 0, 0, 0}, netmask: 128}]] =
-           query("SELECT $1::cidr", [%Postgrex.CIDR{address: {8193, 43981, 0, 0, 0, 0, 0, 0}, netmask: 128}])
+    assert [[%Postgrex.INET{address: {127, 0, 0, 1}, netmask: 32}]] =
+           query("SELECT $1::inet", [%Postgrex.INET{address: {127, 0, 0, 1}, netmask: 32}])
+    assert [[%Postgrex.INET{address: {8193, 43981, 0, 0, 0, 0, 0, 0}, netmask: 128}]] =
+           query("SELECT $1::inet", [%Postgrex.INET{address: {8193, 43981, 0, 0, 0, 0, 0, 0}, netmask: 128}])
+    assert [[%Postgrex.INET{address: {127, 0, 0, 1}, netmask: 8}]] =
+           query("SELECT $1::inet", [%Postgrex.INET{address: {127, 0, 0, 1}, netmask: 8}])
+    assert [[%Postgrex.INET{address: {127, 0, 0, 1}, netmask: 32}]] =
+           query("SELECT $1::cidr", [%Postgrex.INET{address: {127, 0, 0, 1}, netmask: 32}])
+    assert [[%Postgrex.INET{address: {8193, 43981, 0, 0, 0, 0, 0, 0}, netmask: 128}]] =
+           query("SELECT $1::cidr", [%Postgrex.INET{address: {8193, 43981, 0, 0, 0, 0, 0, 0}, netmask: 128}])
+    assert [[%Postgrex.INET{address: {192, 168, 0, 0}, netmask: 16}]] =
+           query("SELECT $1::cidr", [%Postgrex.INET{address: {192, 168, 0, 0}, netmask: 16}])
     assert [[%Postgrex.MACADDR{address: {8, 1, 43, 5, 7, 9}}]] =
            query("SELECT $1::macaddr", [%Postgrex.MACADDR{address: {8, 1, 43, 5, 7, 9}}])
+
+    assert %Postgrex.Error{postgres: %{code: :invalid_binary_representation}} =
+           query("SELECT $1::cidr", [%Postgrex.INET{address: {127, 0, 0, 1}, netmask: 8}])
   end
 
   test "encode bit string", context do