+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Backwards incompatible changes
   * Invoke `encode_to_iodata!` instead of `encode!` in JSON encoder
+  * Remove Postgrex.CIDR and use Postgrex.INET to encode both inet/cidr (as Postgres may perform implicit/explicit casting at any time)
 
 ## v0.13.2 (2017-03-05)